Squashed commit of the following:
commit 1c8afb587583417ce3fc7dfc0bff3d8e59a72e61
Author: Akaza Renn <8340896+AkazaRenn@users.noreply.github.com>
Date:   Sun Jul 7 18:39:12 2024 -0400

    Update play/pause icon color

commit 730f8bf4ceba56272dacf4539563dc559cc212f0
Author: Akaza Renn <8340896+AkazaRenn@users.noreply.github.com>
Date:   Sun Jul 7 18:17:06 2024 -0400

    Update template to latest main

commit c5996fc10c6b1c9eded2b0a174df38721289170d
Author: Akaza Renn <8340896+AkazaRenn@users.noreply.github.com>
Date:   Sun Jul 7 18:01:06 2024 -0400

    Update document

commit 7a0e0d597fe106e2d20acc3ffa53a1a7f42bf23a
Author: Akaza Renn <8340896+AkazaRenn@users.noreply.github.com>
Date:   Sun Jul 7 17:57:15 2024 -0400

    Adjust style

commit 46e84ab535ad69e45e5fa85ef590e0ffa4300ce7
Author: Akaza Renn <8340896+AkazaRenn@users.noreply.github.com>
Date:   Sun Jul 7 17:30:29 2024 -0400

    Button over icon
