+++ .gitignore
@@ -48,3 +48,11 @@ backend/out
 
 # Make sure to ignore any instance of the loader's decky_plugin.py
 decky_plugin.py
+
+# Ignore decky CLI for building plugins
+out
+out/*
+cli/
+cli/*
+cli/decky
+/deck.json
