+++ .vscode/defsettings.json
@@ -5,6 +5,7 @@
     "deckpass"  : "ssap",
     "deckkey"   : "-i ${env:HOME}/.ssh/id_rsa",
     "deckdir"   : "/home/deck",
+    "pluginname": "Pause Games",
     "python.analysis.extraPaths": [
         "./py_modules"
     ]
